+++ src/evloop.c
@@ -8,7 +8,7 @@ static size_t nr_aux_fds = 0;
 struct device device_table[MAX_DEVICES];
 size_t device_table_sz;
 
-static void panic_check(uint8_t code, uint8_t pressed)
+static void panic_check(uint16_t code, uint8_t pressed)
 {
 	static uint8_t enter, backspace, escape;
 	switch (code) {
